Fix enum class conversion errors in MzDataHandler and MzMLSqliteHandler
Add explicit static_cast for ActivationMethod enum class:
- MzDataHandler.cpp: Cast to UInt for writeCVS_ call
- MzMLSqliteHandler.cpp: Cast to int when storing activation_method

+++ src/openms/source/FORMAT/HANDLERS/MzMLSqliteHandler.cpp
@@ -1119,7 +1119,7 @@ namespace OpenMS::Internal
           int activation_method = -1;
           if (!prec.getActivationMethods().empty() )
           {
-            activation_method = *prec.getActivationMethods().begin();
+            activation_method = static_cast<int>(*prec.getActivationMethods().begin());
           }
           String pepseq;
           if (prec.metaValueExists("peptide_sequence"))
@@ -1330,7 +1330,7 @@ namespace OpenMS::Internal
         int activation_method = -1;
         if (!prec.getActivationMethods().empty() )
         {
-          activation_method = *prec.getActivationMethods().begin();
+          activation_method = static_cast<int>(*prec.getActivationMethods().begin());
         }
         String pepseq;
         if (prec.metaValueExists("peptide_sequence"))
